Module antlr_rust::error_strategy

Error handling and recovery

Structs

BailErrorStrategy

This implementation of ANTLRErrorStrategy responds to syntax errors by immediately canceling the parse operation with a ParseCancellationException. The implementation ensures that the ParserRuleContext.exception field is set for all parse tree nodes that were not completed prior to encountering the error.

DefaultErrorStrategy

This is the default implementation of ErrorStrategy used for error reporting and recovery in ANTLR parsers.

ParseCancelledError

ANTLRError::FallThrough Error returned BailErrorStrategy to bail out from parsing

Traits

ErrorStrategy

The interface for defining strategies to deal with syntax errors encountered during a parse by ANTLR-generated parsers. We distinguish between three different kinds of errors: