Struct antlr_rust::input_stream::InputStream

pub struct InputStream<Data: Deref> { /* private fields */ }

Default rust target input stream.

Since Rust uses UTF-8 format which does not support indexing by char, InputStream<&str> has slightly different index behavior in compare to java runtime when there are non-ASCII unicode characters. If you need it to generate exactly the same indexes as Java runtime, you have to use CodePoint8/16/32BitCharStream, which does not use rusts native str type, so it would do additional conversions and allocations along the way.

Implementations

impl<Data: ?Sized + InputData> InputStream<Box<Data>>

pub fn new_owned(data: Box<Data>) -> Self

Creates new InputStream over owned data

impl<'a, Data> InputStream<&'a Data>
where Data: ?Sized + InputData,

pub fn new(data_raw: &'a Data) -> Self

Creates new InputStream over borrowed data

impl<'a, Data: Deref> InputStream<Data>
where Data::Target: InputData,

pub fn reset(&mut self)

Resets input stream to start from the beginning of this slice