Struct antlr_rust::BaseLexer

pub struct BaseLexer<'input, T: LexerRecog<'input, Self> + 'static, Input: CharStream<TF::From>, TF: TokenFactory<'input> = CommonTokenFactory> {
    pub interpreter: Option<Box<LexerATNSimulator>>,
    pub input: Option<Input>,
    pub token_start_char_index: isize,
    pub token_start_line: isize,
    pub token_start_column: isize,
    pub token_type: isize,
    pub token: Option<TF::Tok>,
    pub channel: isize,
    pub mode_stack: Vec<usize>,
    pub mode: usize,
    pub text: Option<<TF::Data as ToOwned>::Owned>,
    /* private fields */
}

Default implementation of Lexer

Public fields in this struct are intended to be used by embedded actions

Fields
interpreter: Option<Box<LexerATNSimulator>>

LexerATNSimulator instance of this lexer

input: Option<Input>

CharStream used by this lexer

token_start_char_index: isizetoken_start_line: isizetoken_start_column: isizetoken_type: isize

Overrides token type emitted by lexer for current token

token: Option<TF::Tok>

Make it Some to override token that is currently being generated by lexer

channel: isize

Channel lexer is currently assigning tokens to

mode_stack: Vec<usize>

stack of modes, which is used for pushMode,popMode lexer actions

mode: usize

Mode lexer is currently in

text: Option<<TF::Data as ToOwned>::Owned>

Make it Some to override text for token that is currently being generated by lexer

Implementations

impl<'input, T, Input, TF> BaseLexer<'input, T, Input, TF>
where T: LexerRecog<'input, Self> + 'static, Input: CharStream<TF::From>, TF: TokenFactory<'input>,

pub fn get_char_index(&self) -> isize

Current position in input stream

pub fn get_text<'a>(&'a self) -> Cow<'a, TF::Data>
where 'input: 'a,

Current token text

pub fn set_text(&mut self, _text: <TF::Data as ToOwned>::Owned)

Used from lexer actions to override text of the token that will be emitted next

pub fn add_error_listener( &mut self, listener: Box<dyn ErrorListener<'input, Self>> )

Add error listener

pub fn remove_error_listeners(&mut self)

Remove and drop all error listeners

pub fn new_base_lexer( input: Input, interpreter: LexerATNSimulator, recog: T, factory: &'input TF ) -> Self

Creates new lexer instance