Trait antlr_rust::token::Token

pub trait Token: Debug + Display {
    type Data: ?Sized + InputData;

    // Required methods
    fn get_token_type(&self) -> isize;
    fn get_text(&self) -> &Self::Data;

    // Provided methods
    fn get_channel(&self) -> isize { ... }
    fn get_start(&self) -> isize { ... }
    fn get_stop(&self) -> isize { ... }
    fn get_line(&self) -> isize { ... }
    fn get_column(&self) -> isize { ... }
    fn set_text(&mut self, _text: <Self::Data as ToOwned>::Owned) { ... }
    fn get_token_index(&self) -> isize { ... }
    fn set_token_index(&self, _v: isize) { ... }
    fn to_owned(&self) -> OwningToken { ... }
}

Implemented by tokens that are produced by a TokenFactory

Required Associated Types

type Data: ?Sized + InputData

Type of the underlying data this token refers to

Required Methods

fn get_token_type(&self) -> isize

fn get_text(&self) -> &Self::Data

Provided Methods

fn get_channel(&self) -> isize

fn get_start(&self) -> isize

fn get_stop(&self) -> isize

fn get_line(&self) -> isize

fn get_column(&self) -> isize

fn set_text(&mut self, _text: <Self::Data as ToOwned>::Owned)

fn get_token_index(&self) -> isize

fn set_token_index(&self, _v: isize)

fn to_owned(&self) -> OwningToken

returns fully owned representation of this token