Type Alias antlr_rust::token::CommonToken

pub type CommonToken<'a> = GenericToken<Cow<'a, str>>;

Most versatile Token that uses Cow to save data Can be used seamlessly switch from owned to zero-copy parsing

Aliased Type

struct CommonToken<'a> {
    pub token_type: isize,
    pub channel: isize,
    pub start: isize,
    pub stop: isize,
    pub token_index: AtomicIsize,
    pub line: isize,
    pub column: isize,
    pub text: Cow<'a, str>,
    pub read_only: bool,
}

Fields
token_type: isizechannel: isizestart: isizestop: isizetoken_index: AtomicIsizeline: isizecolumn: isizetext: Cow<'a, str>read_only: bool